requirements/requirements.txt
Django==2.2.7
djangorestframework==3.10.3
drf-yasg==1.17.0
PyYAML==5.1.2
psycopg2-binary==2.8.4